remove IProviderState interface
