+++ src/Proto.Persistence/IProvider.cs
@@ -4,10 +4,18 @@
 //  </copyright>
 // -----------------------------------------------------------------------
 
+using System;
+using System.Threading.Tasks;
+
 namespace Proto.Persistence
 {
     public interface IProvider
     {
-        IProviderState GetState();
+        Task GetEventsAsync(string actorName, long indexStart, Action<object> callback);
+        Task<(object Snapshot, long Index)> GetSnapshotAsync(string actorName);
+        Task PersistEventAsync(string actorName, long index, object @event);
+        Task PersistSnapshotAsync(string actorName, long index, object snapshot);
+        Task DeleteEventsAsync(string actorName, long inclusiveToIndex);
+        Task DeleteSnapshotsAsync(string actorName, long inclusiveToIndex);
     }
 }
